refactor variable
